BP page for adding results. Fixed avatar display
Instead of doing it all on the events list, which was ugly.
Display of avatar on heads_taken was broken due to modified array added
yesterday.
It's still all a bit of a mess, codewise, but we're getting there
slowly.
Also changed full event list to show all events in descending order